determine file extension from image file content
we now use Image-type to determine the appropriate image file extension
based on the content of the image file itself, and attempting to
extract the extension from the URL is now a fall-back.

I use nprindle's fork of Image-type, which provides for identifying
image type from ByteStrings, since the upstream version of Image-type
currently only supports identifying image type for files.